server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.32.0
transformers>=4.30.0
accelerate>=0.20.0
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=10.0.0
pydantic>=2.0.0
safetensors>=0.3.0
